Trace more metrics

+++ src/omega/trace.cc
@@ -10,17 +10,17 @@ namespace dgl {
 namespace omega {
 
 
-typedef dmlc::moodycamel::ConcurrentQueue<std::tuple<int, std::string, int>> QueueT;
+typedef dmlc::moodycamel::ConcurrentQueue<std::tuple<int, std::string, int64_t>> QueueT;
 
 QueueT queue;
 
-void PutTrace(int batch_id, const std::string& name, int elapsed_micro) {
+void PutTrace(int batch_id, const std::string& name, int64_t elapsed_micro) {
   queue.enqueue(std::make_tuple(batch_id, name, elapsed_micro));
 }
 
-std::vector<std::tuple<int, std::string, int>> GetCppTraces() {
-  std::tuple<int, std::string, int> t;
-  std::vector<std::tuple<int, std::string, int>> ret;
+std::vector<std::tuple<int, std::string, int64_t>> GetCppTraces() {
+  std::tuple<int, std::string, int64_t> t;
+  std::vector<std::tuple<int, std::string, int64_t>> ret;
   while (queue.try_dequeue(t)) {
     ret.emplace_back(std::move(t));
   }
